Update timesheetFunctions.js
Explanation:

getPayPeriodStartDate function:
It takes the currentDate as a parameter.
It calculates the day of the week using currentDate.getDay(), which returns a value from 0 to 6 (0 for Sunday, 1 for Monday, and so on).
It calculates the number of days to subtract from the current date to get to the previous Monday. This is done using the formula (currentDay + 6) % 7. It ensures that we always subtract the correct number of days to reach the previous Monday, even if the current day is already a Monday.
It creates a new Date object called payPeriodStartDate based on the currentDate.
It subtracts the calculated number of days from the payPeriodStartDate using setDate() to get the start date of the pay period.
Finally, it returns the payPeriodStartDate.
getPayPeriodEndDate function:
It takes the startDate (pay period start date) as a parameter.
It creates a new Date object called payPeriodEndDate based on the startDate.
It adds 13 days to the startDate using setDate() to get the end date of the pay period. Since the pay period is a total of 14 days (including the start date), adding 13 days gives us the end date.
Finally, it returns the payPeriodEndDate.
Now, when you call getPayPeriodStartDate(currentDate), it will calculate and return the start date of the pay period based on the current date. The start date will always be a Monday.

Similarly, when you call getPayPeriodEndDate(startDate), it will calculate and return the end date of the pay period based on the provided start date. The end date will always be a Sunday, 13 days after the start date.