Update displayFunctions.js
Here are the main changes and additions made to the displayFunctions.js file:

The stateTimeZones object is now imported from a separate configuration file stateTimeZones.js located in the config directory. This improves maintainability by keeping the time zone data separate from the display functions.
The updateDailyHoursDisplay and updateWeeklyHoursDisplay functions are included in the file. Make sure to call these functions where necessary in your application to update the display of daily and weekly hours.
The updateTimeClockDisplay function has been updated to use a single element timeclockElement to display the clock in, clock out, meal start, and meal end times. It now uses the formatDateTime function to format the date and time values.
The formatDateTime function has been added to format the date and time values consistently throughout the application.
The displayNotification function has been added to display notifications to the user. It takes a message parameter and displays it in the notificationElement with a CSS class 'show' for a specified duration (3 seconds in this example).
The updateEmployeeList function has been added to update the display of the employee list. It takes an array of employees and dynamically generates list items (<li>) for each employee, appending them to the employeeListElement.