Update userData.js
In this updated version of the userData.js file:

The users array remains the same, containing sample user data. Each user object now includes an email property to store the user's email address.
We define and export an asynchronous function called getUserById that takes userId as a parameter.
Inside the getUserById function, we retrieve the user data from the users array based on the provided userId using the Array.find() method.
If a user with the matching userId is found, we return the user object.
If no user is found with the provided userId, we throw an error with an appropriate message.
Please note that this implementation assumes that the user data is stored in memory using the users array. In a real-world application, you would typically retrieve user data from a database or an external API.